Removal of recurring type anotations in `ml`
The duplication of type notations for modules in `ml` is very redundant.
Removing it has allowed specialisation of modules in some
places (notably in Arrows and for Selective). One anotation remains in
the very specific case of Free_applicative, because otherwise the
typechecker cannot consider the equation `'a f = 'a t`.